# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Time Clock API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /time-entries/clock-in:
    post:
      operationId: clockIn
      tags: [timeEntries]
      summary: Clock an employee in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockInInput"
      responses:
        "201":
          description: Clocked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
        "400":
          description: Already clocked in or invalid
  /time-entries/{id}/clock-out:
    patch:
      operationId: clockOut
      tags: [timeEntries]
      summary: Clock an employee out
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockOutInput"
      responses:
        "200":
          description: Clocked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
        "400":
          description: Not clocked in
        "404":
          description: Entry not found
...
  /reports/time-off-balances:
    get:
      operationId: getTimeOffBalances
      tags: [reports]
      summary: Get time off balance for all employees (admin)
      parameters:
        - name: employeeId
          in: query
          required: false
          schema:
            type: integer
        - name: year
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Time off balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeOffBalance"
  /reports/timesheets:
    get:
      operationId: getTimesheetReport
      tags: [reports]
      summary: Get timesheet report for a date range (admin)
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: employeeId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Timesheet report grouped by employee
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeTimesheet"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for the given employee
      parameters:
        - name: employeeId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/unread-count:
    get:
      operationId: getUnreadNotificationCount
      tags: [notifications]
      summary: Get unread notification count for an employee
      parameters:
        - name: employeeId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadNotificationCount"
  /notifications/mark-all-read:
    patch:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read for an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkAllReadBody"
      responses:
        "200":
          description: Updated count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadNotificationCount"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
...
components:
  schemas:
...
    TimeEntry:
      type: object
      required: [id, employeeId, clockIn, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        clockIn:
          type: string
        clockOut:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        totalMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string
    ClockInInput:
      type: object
      required: [employeeId]
      properties:
        employeeId:
          type: integer
        notes:
          type: string
    ClockOutInput:
      type: object
      properties:
        notes:
          type: string
...
    TimesheetEntry:
      type: object
      required: [kind, employeeId]
      properties:
        kind:
          type: string
          enum: [work, time_off]
        id:
          type: ["integer", "null"]
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        clockIn:
          type: ["string", "null"]
        clockOut:
          type: ["string", "null"]
        totalMinutes:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
        timeOffType:
          type: ["string", "null"]
        timeOffRequestId:
          type: ["integer", "null"]
...
    TimeOffTypeBreakdown:
      type: object
      required: [type, usedHours, plannedHours, totalHours]
      properties:
        type:
          type: string
        usedHours:
          type: number
        plannedHours:
          type: number
        totalHours:
          type: number
    TimeOffRequestSummary:
      type: object
      required: [id, startDate, endDate, type, days, hours, status]
      properties:
        id:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        type:
          type: string
        days:
          type: integer
        hours:
          type: number
        status:
          type: string
        notes:
          type: ["string", "null"]
    TimeOffBalance:
      type: object
      required: [employeeId, employeeName, allottedHours, usedHours, plannedHours, remainingHours, usedPlusPlannedHours, year, breakdown, requests, sickTimeAllotmentHours, sickUsedHours, sickPlannedHours, sickRemainingHours]
      properties:
        employeeId:
          type: integer
        employeeName:
          type: string
        department:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        allottedHours:
          type: integer
        usedHours:
          type: number
        plannedHours:
          type: number
        remainingHours:
          type: number
        usedPlusPlannedHours:
          type: number
        sickTimeAllotmentHours:
          type: integer
        sickUsedHours:
          type: number
        sickPlannedHours:
          type: number
        sickRemainingHours:
          type: number
        year:
          type: integer
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/TimeOffTypeBreakdown"
        requests:
          type: array
          items:
            $ref: "#/components/schemas/TimeOffRequestSummary"
...
    Notification:
      type: object
      required: [id, recipientEmployeeId, type, title, message, read, createdAt]
      properties:
        id:
          type: integer
        recipientEmployeeId:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        relatedId:
          type: ["integer", "null"]
        relatedType:
          type: ["string", "null"]
        read:
          type: boolean
        createdAt:
          type: string
    UnreadNotificationCount:
      type: object
      required: [count]
      properties:
        count:
          type: integer
    MarkAllReadBody:
      type: object
      required: [employeeId]
      properties:
        employeeId:
          type: integer
    EmployeeTimesheet:
      type: object
      required: [employeeId, employeeName, totalMinutes, totalTimeOffMinutes, entries]
      properties:
        employeeId:
          type: integer
        employeeName:
          type: string
        department:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        totalMinutes:
          type: integer
        totalTimeOffMinutes:
          type: integer
        entries:
          type: array
          items:
            $ref: "#/components/schemas/TimesheetEntry"