# lib/api-zod/src/generated/types/timeOffBalance.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Time Clock API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TimeOffRequestSummary } from './timeOffRequestSummary';
import type { TimeOffTypeBreakdown } from './timeOffTypeBreakdown';

export interface TimeOffBalance {
  employeeId: number;
  employeeName: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  allottedHours: number;
  usedHours: number;
  plannedHours: number;
  remainingHours: number;
  usedPlusPlannedHours: number;
  sickTimeAllotmentHours: number;
  sickUsedHours: number;
  sickPlannedHours: number;
  sickRemainingHours: number;
  year: number;
  breakdown: TimeOffTypeBreakdown[];
  requests: TimeOffRequestSummary[];
}