# lib/api-zod/src/generated/types/unreadNotificationCount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Time Clock API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UnreadNotificationCount {
  count: number;
}